Feature/cookie and cloud upload (#163)
* Add cookie parameter to /v1/BETA/media/download

* Add cloud_upload parameter to /v1/BETA/media/download

---------

+++ routes/v1/media/download.py
@@ -9,7 +9,7 @@ import uuid
 from services.cloud_storage import upload_file
 from services.authentication import authenticate
 from services.file_management import download_file
-from urllib.parse import quote
+from urllib.parse import quote, urlparse
 
 v1_media_download_bp = Blueprint('v1_media_download', __name__)
 logger = logging.getLogger(__name__)
@@ -22,6 +22,8 @@ logger = logging.getLogger(__name__)
         "media_url": {"type": "string", "format": "uri"},
         "webhook_url": {"type": "string", "format": "uri"},
         "id": {"type": "string"},
+        "cookie": {"type": "string", "description": "Path to cookie file, URL to cookie file, or cookie string in Netscape format"},
+        "cloud_upload": {"type": "boolean"},
         "format": {
             "type": "object",
             "properties": {
@@ -73,6 +75,7 @@ logger = logging.getLogger(__name__)
 @queue_task_wrapper(bypass_queue=False)
 def download_media(job_id, data):
     media_url = data['media_url']
+    cookie = data.get('cookie')
 
     format_options = data.get('format', {})
     audio_options = data.get('audio', {})
@@ -91,7 +94,22 @@ def download_media(job_id, data):
                 'outtmpl': os.path.join(temp_dir, '%(title)s.%(ext)s'),
                 'quiet': True,
                 'no_warnings': True,
+                'download': data.get('cloud_upload', True)
             }
+
+            # Add cookies if provided
+            if cookie:
+                if os.path.isfile(cookie):
+                    ydl_opts['cookiefile'] = cookie
+                elif urlparse(cookie).scheme in ('http', 'https'):
+                    # If cookie is a URL, download it first
+                    ydl_opts['cookiefile'] = download_file(cookie, temp_dir)
+                else:
+                    # If cookie is a string, write it to a temporary file
+                    cookie_file = os.path.join(temp_dir, 'cookies.txt')
+                    with open(cookie_file, 'w') as f:
+                        f.write(cookie)
+                    ydl_opts['cookiefile'] = cookie_file
 
             # Add format options if specified
             if format_options:
@@ -146,19 +164,21 @@ def download_media(job_id, data):
 
             # Download the media
             with yt_dlp.YoutubeDL(ydl_opts) as ydl:
-                info = ydl.extract_info(media_url, download=True)
-                filename = ydl.prepare_filename(info)
+                info = ydl.extract_info(media_url, download=data.get('cloud_upload', True))
                 
-                # Upload to cloud storage
-                cloud_url = upload_file(filename)
-                
-                # Clean up the temporary file
-                os.remove(filename)
+                if not data.get('cloud_upload', True):
+                    media_url = info['url']
+                else:
+                    filename = ydl.prepare_filename(info)
+                    # Upload to cloud storage
+                    media_url = upload_file(filename)
+                    # Clean up the temporary file
+                    os.remove(filename)
 
                 # Prepare response
                 response = {
                     "media": {
-                        "media_url": cloud_url,
+                        "media_url": media_url,
                         "title": info.get('title'),
                         "format_id": info.get('format_id'),
                         "ext": info.get('ext'),
